this will remove timeslots from the nav bar and also refactor the file structure
