+++ App.js
@@ -3,105 +3,33 @@ import * as React from "react";
 import { StyleSheet, Text, View } from "react-native";
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import MaterialCommunityIcons from "react-native-vector-icons/MaterialCommunityIcons";
-import Login from "./Login.jsx";
-import Register from "./Register.jsx";
-import Browse from "./Browse.jsx";
-import TimeSlots from "./TimeSlots.jsx";
-import { UserProvider } from "./context/UserContext";
+import Home from './Screens/Home.jsx';
+import { UserProvider } from './context/UserContext';
+import TimeSlots from './Screens/TimeSlots.jsx';
+
 
 const Stack = createNativeStackNavigator();
-const Tab = createBottomTabNavigator();
+
 
 export default function App() {
   return (
     <UserProvider>
-      <NavigationContainer>
-        <Tab.Navigator>
-          <Tab.Screen
-            name="Login"
-            component={Login}
-            options={{
-              headerShown: false,
-              tabBarStyle: { display: "none" },
-              tabBarLabelStyle: { fontSize: 14 },
-              tabBarStyle: {
-                backgroundColor: "black",
-                paddingTop: 5,
-              },
-              tabBarActiveTintColor: "white",
-              tabBarIcon: ({ size, color }) => (
-                <MaterialCommunityIcons
-                  name="login"
-                  color={"white"}
-                  size={size}
-                />
-              ),
-            }}
-          />
-          <Tab.Screen
-            name="Browse"
-            component={Browse}
-            options={{
-              headerShown: false,
-              tabBarLabelStyle: { fontSize: 14 },
-              tabBarStyle: {
-                backgroundColor: "black",
-                paddingTop: 5,
-              },
-              tabBarActiveTintColor: "white",
-              tabBarIcon: ({ size, color }) => (
-                <MaterialCommunityIcons
-                  name="calendar-month"
-                  color={"white"}
-                  size={size}
-                />
-              ),
-            }}
-          />
-          <Tab.Screen
-            name="Register"
-            component={Register}
-            options={{
-              headerShown: false,
-              tabBarLabelStyle: { fontSize: 14 },
-              tabBarStyle: {
-                backgroundColor: "black",
-                paddingTop: 5,
-              },
-              tabBarActiveTintColor: "white",
-              tabBarIcon: ({ size, color }) => (
-                <MaterialCommunityIcons
-                  name="card-account-details"
-                  color={"white"}
-                  size={size}
-                />
-              ),
-            }}
-          />
-          <Tab.Screen
-            name="TimeSlots"
-            component={TimeSlots}
-            options={{
-              headerShown: false,
-              tabBarLabelStyle: { fontSize: 14 },
-              tabBarStyle: {
-                backgroundColor: "black",
-                paddingTop: 5,
-              },
-              tabBarActiveTintColor: "white",
-              tabBarIcon: ({ size, color }) => (
-                <MaterialCommunityIcons
-                  name="calendar-month"
-                  color={"white"}
-                  size={size}
-                />
-              ),
-            }}
-          />
-        </Tab.Navigator>
+    <NavigationContainer>
+      <Stack.Navigator>
+      <Stack.Screen
+      name="Home" 
+      component={Home}
+      options={{headerShown: false}}
+      />
+       <Stack.Screen
+      name="TimeSlots" 
+      component={TimeSlots}
+      options={{headerShown: false}}
+      />
+      </Stack.Navigator>
       </NavigationContainer>
-    </UserProvider>
+      </UserProvider>
   );
 }
+
+
